pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "isage-libs-intent"
version = "0.1.0.8"
description = "SAGE Libs Intent (L3) - Keyword-based and LLM-based intent classification for conversational AI"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "IntelliStream Team", email = "shuhao_zhang@hust.edu.cn" }]
keywords = [
    "intent-recognition",
    "intent-classification",
    "nlu",
    "conversational-ai",
    "chatbot",
    "llm",
    "keyword-matching",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = ["pydantic>=2.0.0", "typing-extensions>=4.0.0", "openai>=1.52.0,<3.0.0", "anthropic>=0.60.0", "isage-libs>=0.2.0"]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "pytest-cov>=4.0.0", "ruff>=0.8.4", "isage-pypi-publisher>=0.2.0"]

[project.urls]
Homepage = "https://github.com/intellistream/sage-libs-intent"
Repository = "https://github.com/intellistream/sage-libs-intent"
Documentation = "https://github.com/intellistream/sage-libs-intent#readme"
"Bug Tracker" = "https://github.com/intellistream/sage-libs-intent/issues"

[tool.setuptools]
packages = [
    "sage_libs",
    "sage_libs.sage_agentic",
    "sage_libs.sage_agentic.intent",
]
package-dir = { "" = "src" }

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
